feat: add raw events iterator for readers
With the way parse_event is implemented, we cannot differentiate between
an escape sequence and an escape key when using a blocking iterator
(e.g., stdin), this had previously only been tested with a non-blocking
iterator.

Differentiating between an escape sequence and an escape key is now done
in the RawEvents iterator implementation.

RawEvents is implemented to handle the bytes-to-event conversion for
readers, other iterators can build on top of RawEvents for easy handling
of different events.