feat: add alternate buffer support
Most terminals nowadays support two types of screen buffers, the primary
buffer and the alternate buffer.

When enabling the alternate buffer, the terminal will hide the contents
of the primary buffer and enter the alternate buffer. Applications can
write to this buffer as usual, and upon disabling the alternate buffer
(usually done when the application exits), its contents will be cleared,
and the terminal switches back to the primary buffer.

Examples of applications that use the alternate buffer are vim and less.

The alternate buffer is ideal for text editors, as we want to use the
full terminal screen without disrupting the contents of the primary
buffer.