refactor: writing to stdout
Using the print! macro does come with a very small chance of panicking
if writing to stdout fails, but I believe this chance compared to the
benefits is small enough to not worry about it.

We still need to keep the reference to the RawTerminal and
AlternateBuffer alive so they don't run their destructors.