refactor: size structs
Before, we were using the terminal size struct for other things than the
terminal, this puts a limit on things as the width and height are only
u16. For example, if the file had more lines than an u16 allows, the
cursor would not be able to reach the bottom.