feat: add panic hook
Since the alternate buffer was implemented, you have been unable to view
the panic information, as the information gets printed to the alternate
buffer, which gets disabled when the AlternateBuffer struct gets
dropped.

This has been quite annoying for development.

These changes fix this issue by using a panic hook to
disable the alternate buffer and raw mode *before* printing the panic
information.